avocado.core: Merge avocado.core.log into avocado.core.output
Those two are interconnected, let's keep them in one file instead of
local imports.
